requirements.txt
# This file lists the dependencies for the project. You can install them using pip:
Pillow
openai==1.30.1
python-dotenv
httpx==0.27.0